refactor(Lotto): WinningException 클래스 대신 Lotto 클래스에서 예외처리
