+++ docs/README.md
@@ -52,6 +52,7 @@
   - [x] 로또 구입 금액이 정수가 아닌 경우
   - [x] 로또 구입 금액이 1000원으로 나누어 떨어지지 않는 경우
 
+  - [x] 당첨 번호가 정수가 아닌 경우
   - [x] 당첨 번호가 6개가 아닌 경우
   - [x] 당첨 번호가 1 - 45 사이의 수가 아닌 경우
   - [x] 당첨 번호가 중복되는 경우
